requirements.txt
fastapi==0.115.6
numpy==1.24.3
pillow==11.0.0
joblib==1.4.2
tensorflow==2.13.0